feat: 기초적인 CRUD 구현, 테스트X

+++ src/main/java/com/billit/loan_service/dto/LoanRequest.java
@@ -9,7 +9,7 @@ import java.math.BigDecimal;
 @Getter
 @AllArgsConstructor
 public class LoanRequest {
-    private Long userBorrowId;
+    private Integer userBorrowId;
     private BigDecimal loanAmount;
     private Integer term;
 }
